update a11y and visual tests
