+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-dark-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:658b1f9fc5f2c4f22656b7d58a8dd61d7443eac0cc9e6b90d85cc73e1d15c9ab
+size 2196
