+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-dark-horizontal-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:fac3afe4bd548c0275f9febd958082236a74a0af07a5aae77c326988ab42ebaf
+size 1197
