+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-dark-horizontal.ico
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:473f01abf0ddfe1f0e9f90a17f38999469bc26f0b50f5418875a7fcc124bf03c
-size 919
+oid sha256:5f62aa3bfc18420db38332f02607090a11d6b9131ad2201072fe942123d20c54
+size 1608
