+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-dark-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:348d7f8e8e4447180d11802ac48bf495a61a7197c8749686dcc887f15f5fbf32
+size 1554
