+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-highcontrast-dark-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:908c59a342da1b2779ef207259975371d8ca222de8dfe2c047eb47b830963e6c
+size 2279
