+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-highcontrast-dark-horizontal-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:fda5c220d32a6915d5e5342c2ed4f26eed36d51f27e09f3e939497de2f99edfd
+size 1201
