+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-highcontrast-dark-horizontal.ico
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6523c9258a86e92316a8294ac15608e9afa049a946ec55dadf6789fff0622e84
-size 949
+oid sha256:fd8f1e90b6f3bbec621c8174e44a2591330ada56f4476e607c4a75ba84b60dd9
+size 1616
