+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-highcontrast-dark-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:10f150102a6a9ef91958d0977ca045f172e6bb00b2bac471944372c6d5a790a4
+size 1608
