+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-highcontrast-light-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:51cc72a5d0557f76ae89d289a6c8fc21f368e68a0ce69779a73572a058106b5f
+size 2193
