+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-highcontrast-light-horizontal-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6f4c0ca6921e0d835e434165ce8a29035355db530cc990eea5ef65dc01016e22
+size 1235
