+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-light-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d3715ae1f76a4c63d797b9b33cbfe04fb5eee6cd7ffd3366461143fa10234589
+size 2201
