+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-light-horizontal.ico
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:92de5779d55fad16ae1987fe7d147901e9f616dcad38520bace7d0faeaad3a79
-size 908
+oid sha256:e1b5ab8f70790a1edd91bc6ec80e1a82315cda0ff964b735857c6850e0ea3396
+size 1745
