+++ packages/stacks-classic/screenshots/Chromium/baseline/s-vote-light-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1d99deb5818061f82e07627d1fe5f3c9e4f08fa6e5e768bc3fd2ee2dc6e5054d
+size 1590
