+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-dark-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a0b25279508fa575d04e737fc5ef9377c40a80a2dcab085a7d53b5f0cd6c274c
+size 2888
