+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-dark-horizontal-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:41f17ced5b1db6302de8302aa256cf927d68482ad602999c5fb610665d3b16e3
+size 1283
