+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-dark-horizontal.ico
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:468d6a9907560dbbe79100d245736cd3fd5306fe6526cbd86cc50e5737352ff8
-size 1038
+oid sha256:6ced3650cf1b7a22cd9d64620eb46ac2f8ae818db19a17bd9131b83e7bac261a
+size 1731
