+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-dark-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:dd00a42a5f711a539848bb002413e9e094242bf4a6b5bb9408c9774fb1eee3ba
+size 2024
