+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-highcontrast-dark-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ed6904263e0edd6b326ae9eb77604d5ee021f40ad2f7149660462c49df67045c
+size 2972
