+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-highcontrast-dark-horizontal-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:eb65fbb291962d9d08f14352d67705da9b83a225641c71ded4c2d1cc90fd4602
+size 1370
