+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-highcontrast-dark-horizontal.ico
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1fc1dc9fe8454574d535d966126a8db424291c76b42f118b7c53e1a04452a461
-size 1098
+oid sha256:c5a622679c543664b86f8079fd7a90beecc2aeda919b840dd7094cb814540f5d
+size 1804
