+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-highcontrast-dark-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6f62ae92af288259d5b2211d3c01402463a877f4778e5e98709a9ed96bfcf414
+size 2077
