+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-highcontrast-light-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:201feeda026d1fee79f7f41616cd2eeec716ddd15195b51924f9444acd8b08e8
+size 2907
