+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-highcontrast-light-horizontal-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:8c3b4c717654dc41ea373cbff51adf0c19d7a859193420f2b0d08b859ddaa8dc
+size 1412
