+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-highcontrast-light-horizontal.ico
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:eb5e5e47c6708e628a4a681051e2fa00a08b7a21d395f20fa14f34ed8f3ebf78
-size 1069
+oid sha256:5f66750766e4e63e7e0bc56a97cbe645d32ce741cb80af82409bf40ce399bc39
+size 1965
