+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-light-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:8f299e88dfe067a3ec7a84ff4528af7b83e11c13584e3398679a027a2a8232dd
+size 2912
