+++ packages/stacks-classic/screenshots/Firefox/baseline/s-vote-light-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d081b8872cea0b0683bde70e94fb4168c8b033dff475ecdfd47c05ae92b001c2
+size 2126
