+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-dark-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7d50cb212af126d30b59aaebc5c0a9a9b1de897fcd2f08907e7b09a94362efca
+size 2568
