+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-dark-horizontal-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:45490274184d8a2a923149e13ebd7b489030a265f4531b93d29d4060a96b55d7
+size 1418
