+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-dark-horizontal.ico
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e80f456f6966b741953340f66fa45ff86f768cd41939df98fff6f44aa88690e5
-size 1085
+oid sha256:f7396e043067186c77b9cc768a4ea946d1f19a726c3999d10c802114c93360c2
+size 1963
