+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-dark-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:61357508c36ad9eb9e740d688d5ac5b714a9586aa9109d20c3012191a2415e28
+size 1892
