+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-highcontrast-dark-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:8751d86402a2d6eac1b78acdd07d6c66d0f92d535e0e95992fde1ad084f36110
+size 2614
