+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-highcontrast-dark-horizontal-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f224d86fbbe73f20487830c9e4162fe9ba827c8bd275493979491521e225e575
+size 1445
