+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-highcontrast-dark-horizontal.ico
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:105ab4e3137f400c5a8392a5a84e07df159d2d47ed440e20d79e4654666c1b9e
-size 1101
+oid sha256:20121295ec79a0fc91d78638ad5d55aba6645f30ab6ab57efacd8307f280bd18
+size 1980
