+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-highcontrast-dark-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b8ba4ba9c60bded3bfbb958503e7472d9c36b2b2c22223dc7705bad9ebeed08a
+size 1920
