+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-highcontrast-light-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:e7089b89385e9e65f296bb03016c1cf3b664cf54ccc898f3ab1caa31dd450db5
+size 2541
