+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-highcontrast-light-horizontal-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:28a0bb87e5b4fe3f79292567350cc7d1d4d3f5193ee022219a4f3be637c748e5
+size 1394
