+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-light-expanded-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:3350a3f2bc49783dd938909b560738910a23b0fb7399760875f0c2ee4e3e25ad
+size 2530
