+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-light-horizontal.ico
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:de5a36442f62eefe18c7ce96fa5a0b3847b93dae20a4e142c58318a6526b42b2
-size 1067
+oid sha256:dd7896f616e268fe2ecf65afbe5f1ea81073f310248bdfc9b8f785508bfd8005
+size 1968
