+++ packages/stacks-classic/screenshots/Webkit/baseline/s-vote-light-upvoteOnly.ico
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:241a8d977ca39f5778f7adcda1e35cc58bada53f3030612297423e2d2db633e1
+size 1889
